# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Video-igre",
    "description": "API koji pruža pristup resursima koji se unutar skupa podataka definiranog na https://github.com/Jurica-Laljak/Video-igre",
    "version": "1.0.0",
    "contact": {
      "name": "Jurica Laljak",
      "url": "https://github.com/Jurica-Laljak",
      "email": "jurica.laljak@fer.unizg.hr"
    },
    "license": {
      "name": "GNU General Public License",
      "url": "https://www.gnu.org/licenses/gpl-3.0.txt"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5173/api/v1"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message",
          "response"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "response": {
            "type": "object"
          }
        }
      },
      "Dlc": {
        "type": "object",
        "required": [
          "name",
          "price",
          "release_date"
        ],
        "items": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            },
            "release_date": {
              "type": "string"
            }
          }
        }
      },
      "Dlc_extended": {
        "type": "object",
        "required": [
          "name",
          "price",
          "release_date",
          "game"
        ],
        "items": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            },
            "release_date": {
              "type": "string"
            },
            "game": {
              "type": "string"
            }
          }
        }
      },
      "Video_game": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "developer": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "platforms": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "metascore": {
            "type": "integer"
          },
          "has_singleplayer": {
            "type": "boolean"
          },
          "has_multiplayer": {
            "type": "boolean"
          },
          "dlc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dlc"
            }
          }
        }
      },
      "Video_game_complete": {
        "type": "object",
        "required": [
          "name",
          "release_date",
          "developer",
          "publisher",
          "platforms",
          "genre",
          "price",
          "metascore",
          "has_singleplayer",
          "has_multiplayer",
          "dlc"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "developer": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "platforms": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "metascore": {
            "type": "integer"
          },
          "has_singleplayer": {
            "type": "boolean"
          },
          "has_multiplayer": {
            "type": "boolean"
          },
          "dlc": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dlc"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/openapi": {
      "get": {
        "description": "Dohvaća OpenAPI 3.0.3 specifikaciju koja opisuje ovo sučelje",
        "operationId": "specificationGet",
        "responses": {
          "200": {
            "description": "Specifikacija uspješno dohvaćena",
            "content": {
              "application/json": {}
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju specifikacije",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/collection": {
      "get": {
        "description": "Dohvaća cjelokupnu kolekciju videoigara",
        "operationId": "collectionGet",
        "responses": {
          "200": {
            "description": "Kolekcija uspješno dohvaćena",
            "content": {
              "application/json": {}
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju kolekcije",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/resource": {
      "post": {
        "description": "Dodaje zapis o video igri u kolekciju",
        "operationId": "resourcePost",
        "requestBody": {
          "description": "Video igra koja će dodati u kolekciju",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Video_game_complete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zapis o video igri uspješno dodan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video_game_complete"
                }
              }
            }
          },
          "400": {
            "description": "Objekt nije ispravo definiran",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video_game_complete"
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u dodavanju zapisa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/resource/{id}": {
      "get": {
        "description": "Dohvaća zapis resursa s danim ID-om",
        "operationId": "resourceGet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id zapisa",
            "required": true,
            "content": {
              "text/html": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resurs uspješno dohvaćen",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Ne postoji resurs s danim ID-om",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju kolekcije",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Ažurira vrijednosti atributa resursa s danim ID-om",
        "operationId": "resourcePut",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id zapisa",
            "required": true,
            "content": {
              "text/html": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "requestBody": {
          "description": "Popis atributa čija će se vrijednost ažurirati i pripadajućih vrijednosti",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Video_game"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vrijednosti atributa resursa uspješno ažurirane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "object": {
                      "$ref": "#/components/schemas/Video_game_complete"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Objekt nije ispravno definiran",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video_game_complete"
                }
              }
            }
          },
          "404": {
            "description": "Ne postoji resurs s danim ID-om",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u ažuriranju resursa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Briše resurs s danim ID-om",
        "operationId": "resourceDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id zapisa",
            "required": true,
            "content": {
              "text/html": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resurs uspješno obrisan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video_game_complete"
                }
              }
            }
          },
          "404": {
            "description": "Ne postoji resurs s danim ID-om",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u brisanju resursa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/genre": {
      "get": {
        "description": "Dohvaća popis žanrova svih videoigara koje se nalaze u kolekciji",
        "operationId": "genreGet",
        "responses": {
          "200": {
            "description": "Žanrovi uspješno dohvaćeni",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju žanrova",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/ids": {
      "get": {
        "description": "Dohvaća popis svih id-ova i pripadajućih naziva igara",
        "operationId": "highratedGet",
        "responses": {
          "200": {
            "description": "Popis svih id-ova dohvaćen",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju popisa id-ova",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/dlc": {
      "get": {
        "description": "Dohvaća popis svih DLC-ova",
        "operationId": "dlcGet",
        "responses": {
          "200": {
            "description": "DLC-ovi uspješno dohvaćeni",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dlc_extended"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Pogreška u dohvaćanju DLC-ova",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}